clean thread.c

+++ Aufgabe1/vorgabe/include/thread.h
@@ -27,9 +27,9 @@ struct thcStruct{
 void initThreadArray();
 void cpyStacktoTHC();
 void createThread(void (*func)(void *), const void*, uint32_t);
-uint16_t getRunningThread();
+int getRunningThread();
 int16_t threadWaitingForChar();
-uint16_t getDeadThread();
+int getDeadThread();
 void endThread();
 void killThread(uint16_t);
 void saveContext(uint16_t, void*);
